+++ lld/ELF/Writer.cpp
@@ -1767,6 +1767,10 @@ template <class ELFT> void Writer<ELFT>::finalizeAddressDependentContent() {
              osec->name + " is not a multiple of alignment (" +
              Twine(osec->addralign) + ")");
     }
+
+  // Sizes are no longer allowed to grow, so all allowable spills have been
+  // taken. Remove any leftover potential spills.
+  script->eraseSpillSections();
 }
 
 // If Input Sections have been shrunk (basic block sections) then