+++ lld/docs/ReleaseNotes.rst
@@ -31,6 +31,12 @@ ELF Improvements
   relocatable object file. (`D146778 <https://reviews.llvm.org/D146778>`_)
 * common-page-size can now be larger than the system page-size.
   (`#57618 <https://github.com/llvm/llvm-project/issues/57618>`_)
+* ``--enable-non-contiguous-regions`` option allows automatically packing input
+  sections into memory regions by automatically spilling to later matches if a
+  region would overflow. This reduces the toil of manually packing regions
+  (typical for embedded). It also makes full LTO feasible in such cases, since
+  IR merging currently prevents the linker script from referring to input
+  files.
 
 Breaking changes
 ----------------